Refactor ToolCommand implementation
