+++ cli/ballerina-cli/src/main/java/io/ballerina/cli/cmd/Constants.java
@@ -43,6 +43,12 @@ public final class Constants {
     public static final String GRAPH_COMMAND = "graph";
     public static final String DEPRECATE_COMMAND = "deprecate";
     public static final String TOOL_COMMAND = "tool";
+    public static final String TOOL_PULL_COMMAND = "pull";
+    public static final String TOOL_LIST_COMMAND = "list";
+    public static final String TOOL_USE_COMMAND = "use";
+    public static final String TOOL_SEARCH_COMMAND = "search";
+    public static final String TOOL_REMOVE_COMMAND = "remove";
+    public static final String TOOL_UPDATE_COMMAND = "update";
     public static final String SEMVER_COMMAND = "semver";
     public static final String DIST_COMMAND = "dist";
     public static final String UPDATE_COMMAND = "update";
