+++ cli/ballerina-cli/src/test/java/io/ballerina/cli/cmd/BaseCommandTest.java
@@ -22,7 +22,9 @@ import io.ballerina.cli.BLauncherCmd;
 import io.ballerina.cli.launcher.BLauncherException;
 import io.ballerina.projects.util.ProjectConstants;
 import io.ballerina.projects.util.ProjectUtils;
+import org.apache.commons.io.FileUtils;
 import org.ballerinalang.compiler.BLangCompilerException;
+import org.ballerinalang.test.BCompileUtil;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.AfterMethod;
@@ -70,12 +72,45 @@ public abstract class BaseCommandTest {
 
     @BeforeSuite
     public void copyBalTools() throws IOException {
+        Path testResourcesSrc = Paths.get("src/test/resources/test-resources/buildToolResources/tools");
+        Path testResources = Files.createTempDirectory("b7a-cmd-test-" + System.nanoTime());
         // copy the bal-tools.toml to <user-home>/.ballerina/.config/
         Path balToolsTomlSrcPath = Paths.get("src/test/resources/test-resources/" +
                 "buildToolResources/tools/bal-tools.toml");
         Path balToolsTomlDstPath = testDotBallerina.resolve(".config");
         Files.createDirectories(balToolsTomlDstPath);
         Files.copy(balToolsTomlSrcPath, balToolsTomlDstPath.resolve("bal-tools.toml"));
+
+
+        FileUtils.copyDirectory(testResourcesSrc.toFile(), testResources.toFile());
+        // compile and cache test build tools
+        String sampleBuildToolJar = "sample-build-tool-1.0.0.jar";
+        Path sampleBuildToolJarPath = Paths.get("build/tool-libs").resolve(sampleBuildToolJar);
+        Path destPath = testResources.resolve("sample-build-tool-pkg")
+                .resolve("lib").resolve(sampleBuildToolJar);
+        Files.createDirectories(destPath.getParent());
+        Files.copy(sampleBuildToolJarPath, destPath);
+        destPath = testResources.resolve("dummy-tool-pkg-higher-dist")
+                .resolve("lib").resolve(sampleBuildToolJar);
+        Files.createDirectories(destPath.getParent());
+        Files.copy(sampleBuildToolJarPath, destPath);
+
+        BCompileUtil.compileAndCacheBala(
+                testResources.resolve("dummy-tool-pkg").toString(), testCentralRepoCache);
+        BCompileUtil.compileAndCacheBala(
+                testResources.resolve("hidden-cmd-tool-pkg").toString(), testCentralRepoCache);
+        BCompileUtil.compileAndCacheBala(
+                testResources.resolve("missing-interface-tool-pkg").toString(), testCentralRepoCache);
+        BCompileUtil.compileAndCacheBala(
+                testResources.resolve("no-options-tool-pkg").toString(), testCentralRepoCache);
+        BCompileUtil.compileAndCacheBala(
+                testResources.resolve("invalid-name-tool-pkg").toString(), testCentralRepoCache);
+        BCompileUtil.compileAndCacheBala(
+                testResources.resolve("dummy-tool-pkg-higher-dist").toString(), testCentralRepoCache);
+        BCompileUtil.compileAndCacheBala(
+                testResources.resolve("ballerina-generate-file").toString(), testCentralRepoCache);
+        BCompileUtil.compileAndCacheBala(
+                testResources.resolve("sample-build-tool-pkg").toString(), testCentralRepoCache);
     }
 
     @BeforeMethod
@@ -162,8 +197,12 @@ public abstract class BaseCommandTest {
 
     @AfterMethod (alwaysRun = true)
     public void afterMethod() throws IOException {
-        console.close();
-        printStream.close();
+        if (console != null) {
+            console.close();
+        }
+        if (printStream != null) {
+            printStream.close();
+        }
     }
 
     @AfterClass (alwaysRun = true)
