+++ cli/ballerina-cli/src/test/java/io/ballerina/cli/cmd/ToolCommandTest.java
@@ -63,12 +63,12 @@ public class ToolCommandTest extends BaseCommandTest {
 
     @Test(description = "Pull a tool from local repository")
     public void testPullToolFromLocal() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("pull", "luhee:1.1.0", "--repository=local");
-        toolCommand.execute();
+        ToolPullCommand toolPullCommand = new ToolPullCommand(printStream, printStream, false);
+        new CommandLine(toolPullCommand).parseArgs("luhee:1.1.0", "--repository=local");
+        toolPullCommand.execute();
 
         String buildLog = readOutput(true);
-        Assert.assertEquals(buildLog.replace("\r", ""), "tool 'luhee:1.1.0' successfully set as the active version.\n");
+        Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-pull-success.txt"));
         Assert.assertTrue(Files.exists(BalToolsUtil.BAL_TOOLS_TOML_PATH));
         BalToolsToml balToolsToml = BalToolsToml.from(BalToolsUtil.BAL_TOOLS_TOML_PATH);
         BalToolsManifest balToolsManifest = BalToolsManifestBuilder.from(balToolsToml).build();
@@ -87,12 +87,11 @@ public class ToolCommandTest extends BaseCommandTest {
 
     @Test(description = "Switch active version from local")
     public void testUseToolFromLocal() throws IOException {
-
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("pull", "luhee:1.1.0", "--repository=local");
-        toolCommand.execute();
-        new CommandLine(toolCommand).parseArgs("pull", "luhee:1.2.0", "--repository=local");
-        toolCommand.execute();
+        ToolPullCommand toolPullCommand = new ToolPullCommand(printStream, printStream, false);
+        new CommandLine(toolPullCommand).parseArgs("luhee:1.1.0", "--repository=local");
+        toolPullCommand.execute();
+        new CommandLine(toolPullCommand).parseArgs("luhee:1.2.0", "--repository=local");
+        toolPullCommand.execute();
 
         String buildLog = readOutput(true);
         Assert.assertTrue(buildLog.replace("\r", "")
@@ -112,9 +111,9 @@ public class ToolCommandTest extends BaseCommandTest {
         Assert.assertEquals(tool.get().org(), "gayaldassanayake");
         Assert.assertEquals(tool.get().name(), "tool_gayal");
 
-        toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("use", "luhee:1.1.0", "--repository=local");
-        toolCommand.execute();
+        ToolUseCommand toolUseCommand = new ToolUseCommand(printStream, printStream, false);
+        new CommandLine(toolUseCommand).parseArgs("luhee:1.1.0", "--repository=local");
+        toolUseCommand.execute();
         buildLog = readOutput(true);
         Assert.assertTrue(buildLog.replace("\r", "")
                 .contains("tool 'luhee:1.1.0' successfully set as the active version.\n"));
@@ -134,11 +133,11 @@ public class ToolCommandTest extends BaseCommandTest {
 
     @Test(description = "Remove a tool from local repository")
     public void testRemoveToolFromLocal() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("pull", "luhee:1.1.0", "--repository=local");
-        toolCommand.execute();
-        new CommandLine(toolCommand).parseArgs("pull", "luhee:1.2.0", "--repository=local");
-        toolCommand.execute();
+        ToolPullCommand toolPullCommand = new ToolPullCommand(printStream, printStream, false);
+        new CommandLine(toolPullCommand).parseArgs("luhee:1.1.0", "--repository=local");
+        toolPullCommand.execute();
+        new CommandLine(toolPullCommand).parseArgs("luhee:1.2.0", "--repository=local");
+        toolPullCommand.execute();
 
         String buildLog = readOutput(true);
         Assert.assertTrue(buildLog.replace("\r", "")
@@ -158,9 +157,9 @@ public class ToolCommandTest extends BaseCommandTest {
         Assert.assertEquals(tool.get().org(), "gayaldassanayake");
         Assert.assertEquals(tool.get().name(), "tool_gayal");
 
-        toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("use", "luhee:1.1.0", "--repository=local");
-        toolCommand.execute();
+        ToolUseCommand toolUseCommand = new ToolUseCommand(printStream, printStream, false);
+        new CommandLine(toolUseCommand).parseArgs("luhee:1.1.0", "--repository=local");
+        toolUseCommand.execute();
         buildLog = readOutput(true);
         Assert.assertTrue(buildLog.replace("\r", "")
                 .contains("tool 'luhee:1.1.0' successfully set as the active version.\n"));
@@ -176,9 +175,9 @@ public class ToolCommandTest extends BaseCommandTest {
         Assert.assertEquals(tool.get().org(), "gayaldassanayake");
         Assert.assertEquals(tool.get().name(), "tool_gayal");
 
-        toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("remove", "luhee:1.2.0", "--repository=local");
-        toolCommand.execute();
+        ToolRemoveCommand toolRemoveCommand = new ToolRemoveCommand(printStream, printStream, false);
+        new CommandLine(toolRemoveCommand).parseArgs("luhee:1.2.0", "--repository=local");
+        toolRemoveCommand.execute();
         buildLog = readOutput(true);
         Assert.assertTrue(buildLog.replace("\r", "")
                 .contains("tool 'luhee:1.2.0' successfully removed.\n"));
@@ -194,210 +193,163 @@ public class ToolCommandTest extends BaseCommandTest {
     public void testToolCommandWithHelpFlag() throws IOException {
         String expected = getOutput("tool-help.txt");
 
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
+        ToolCommand toolCommand = new ToolCommand(printStream);
         new CommandLine(toolCommand).parseArgs("--help");
         toolCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replace("\r", ""), expected);
 
-        toolCommand = new ToolCommand(printStream, printStream, false);
+        toolCommand = new ToolCommand(printStream);
         new CommandLine(toolCommand).parseArgs("-h");
+        toolCommand.execute();
+        buildLog = readOutput(true);
+        Assert.assertEquals(buildLog.replace("\r", ""), expected);
+
+        toolCommand = new ToolCommand(printStream);
+        new CommandLine(toolCommand).parseArgs();
         toolCommand.execute();
         buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replace("\r", ""), expected);
     }
 
-    @Test(description = "Test tool command with no arguments")
-    public void testToolCommandWithNoArgs() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs();
-        toolCommand.execute();
-        String buildLog = readOutput(true);
-        Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-with-no-args.txt"));
-    }
-
-    @Test(description = "Test tool command with invalid sub command")
-    public void testToolCommandWithInvalidSubCommand() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("invalid-cmd");
-        toolCommand.execute();
-        String buildLog = readOutput(true);
-        Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-with-invalid-sub-command.txt"));
-    }
-
     @Test(description = "Test tool pull sub-command with no arguments")
     public void testToolPullSubCommandWithNoArgs() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("pull");
-        toolCommand.execute();
+        ToolPullCommand toolPullCommand = new ToolPullCommand(printStream, printStream, false);
+        toolPullCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-pull-with-no-args.txt"));
     }
 
     @Test(description = "Test tool pull sub-command with too many arguments")
     public void testToolPullSubCommandWithTooManyArgs() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("pull", "arg1", "arg2");
-        toolCommand.execute();
+        ToolPullCommand toolPullCommand = new ToolPullCommand(printStream, printStream, false);
+        new CommandLine(toolPullCommand).parseArgs("arg1", "arg2");
+        toolPullCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-pull-with-too-many-args.txt"));
     }
 
     @Test(description = "Test tool pull sub-command with invalid argument format")
     public void testToolPullSubCommandWithInvalidArgFormat() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("pull", "id:1.0.1:extra");
-        toolCommand.execute();
+        ToolPullCommand toolPullCommand = new ToolPullCommand(printStream, printStream, false);
+        new CommandLine(toolPullCommand).parseArgs("id:1.0.1:extra");
+        toolPullCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-pull-with-invalid-tool-id.txt"));
     }
 
     @Test(dataProvider = "invalidToolIds", description = "Test tool pull sub-command with invalid argument format")
     public void testToolPullSubCommandWithInvalidToolId(String toolId) throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("pull", toolId);
-        toolCommand.execute();
+        ToolPullCommand toolPullCommand = new ToolPullCommand(printStream, printStream, false);
+        new CommandLine(toolPullCommand).parseArgs(toolId);
+        toolPullCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-pull-with-invalid-tool-id.txt"));
     }
 
     @Test(description = "Test tool pull sub-command with invalid tool version")
     public void testToolPullSubCommandWithInvalidToolVersion() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("pull", "tool_id:1.1");
-        toolCommand.execute();
+        ToolPullCommand toolPullCommand = new ToolPullCommand(printStream, printStream, false);
+        new CommandLine(toolPullCommand).parseArgs("tool_id:1.1");
+        toolPullCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-pull-with-invalid-tool-version.txt"));
     }
 
-    @Test(description = "Test tool list sub-command with arguments")
-    public void testToolListSubCommandWithArgs() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("list", "arg");
-        toolCommand.execute();
-        String buildLog = readOutput(true);
-        Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-list-with-invalid-arg.txt"));
-
-        toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("list", "arg1", "arg2");
-        toolCommand.execute();
-        buildLog = readOutput(true);
-        Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-list-with-too-many-args.txt"));
-    }
-
     @Test(description = "Test tool remove with more than one argument")
     public void testToolRemoveSubCommandWithTooManyArgs() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("remove", "arg1", "arg2");
-        toolCommand.execute();
+        ToolRemoveCommand toolRemoveCommand = new ToolRemoveCommand(printStream, printStream, false);
+        new CommandLine(toolRemoveCommand).parseArgs("arg1", "arg2");
+        toolRemoveCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-remove-with-too-many-args.txt"));
     }
 
     @Test(description = "Test tool remove with more than no arguments")
     public void testToolRemoveSubCommandWithNoArgs() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("remove");
-        toolCommand.execute();
+        ToolRemoveCommand toolRemoveCommand = new ToolRemoveCommand(printStream, printStream, false);
+        toolRemoveCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-remove-with-no-args.txt"));
     }
 
     @Test(description = "Test tool remove sub-command with invalid argument format")
     public void testToolRemoveSubCommandWithInvalidArgFormat() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("remove", "id:1.0.1:extra");
-        toolCommand.execute();
+        ToolRemoveCommand toolRemoveCommand = new ToolRemoveCommand(printStream, printStream, false);
+        new CommandLine(toolRemoveCommand).parseArgs("id:1.0.1:extra");
+        toolRemoveCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-remove-with-invalid-tool-id.txt"));
     }
 
     @Test(dataProvider = "invalidToolIds", description = "Test tool remove sub-command with invalid argument format")
     public void testToolRemoveSubCommandWithInvalidToolId(String toolId) throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("remove", toolId);
-        toolCommand.execute();
+        ToolRemoveCommand toolRemoveCommand = new ToolRemoveCommand(printStream, printStream, false);
+        new CommandLine(toolRemoveCommand).parseArgs(toolId);
+        toolRemoveCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-remove-with-invalid-tool-id.txt"));
     }
 
     @Test(description = "Test tool pull sub-command with invalid tool version")
     public void testToolRemoveSubCommandWithInvalidToolVersion() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("remove", "tool_id:1.1");
-        toolCommand.execute();
+        ToolRemoveCommand toolRemoveCommand = new ToolRemoveCommand(printStream, printStream, false);
+        new CommandLine(toolRemoveCommand).parseArgs("tool_id:1.1");
+        toolRemoveCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-remove-with-invalid-tool-version.txt"));
     }
 
     @Test(description = "Test tool search with more than one argument")
     public void testToolSearchSubCommandWithTooManyArgs() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("search", "arg1", "arg2");
-        toolCommand.execute();
+        ToolSearchCommand toolSearchCommand = new ToolSearchCommand(printStream, printStream, false);
+        new CommandLine(toolSearchCommand).parseArgs("arg1", "arg2");
+        toolSearchCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-search-with-too-many-args.txt"));
     }
 
-    @Test(description = "Test tool search with more than no arguments")
-    public void testToolSearchSubCommandWithNoArgs() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("search");
-        toolCommand.execute();
-        String buildLog = readOutput(true);
-        Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-search-with-no-args.txt"));
-    }
-
     @Test(description = "Test tool use with more than one argument")
     public void testToolUseSubCommandWithTooManyArgs() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("use", "arg1", "arg2");
-        toolCommand.execute();
+        ToolUseCommand toolUseCommand = new ToolUseCommand(printStream, printStream, false);
+        new CommandLine(toolUseCommand).parseArgs("arg1", "arg2");
+        toolUseCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-use-with-too-many-args.txt"));
     }
 
-    @Test(description = "Test tool use with more than no arguments")
-    public void testToolUseSubCommandWithNoArgs() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("use");
-        toolCommand.execute();
-        String buildLog = readOutput(true);
-        Assert.assertEquals(buildLog.replace("\r", ""), getOutput("tool-use-with-no-args.txt"));
-    }
-
     @Test(description = "Test tool use sub-command with invalid argument format")
     public void testToolUseSubCommandWithInvalidArgFormat() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("use", "id:1.0.1:extra");
-        toolCommand.execute();
+        ToolUseCommand toolUseCommand = new ToolUseCommand(printStream, printStream, false);
+        new CommandLine(toolUseCommand).parseArgs("id:1.0.1:extra");
+        toolUseCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replaceAll("\r", ""), getOutput("tool-use-with-invalid-tool-id.txt"));
     }
 
     @Test(description = "Test tool use sub-command with no version")
     public void testToolUseSubCommandWithNoVersion() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("use", "id");
-        toolCommand.execute();
+        ToolUseCommand toolUseCommand = new ToolUseCommand(printStream, printStream, false);
+        new CommandLine(toolUseCommand).parseArgs("id");
+        toolUseCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replaceAll("\r", ""), getOutput("tool-use-with-invalid-tool-id.txt"));
     }
 
     @Test(dataProvider = "invalidToolIds", description = "Test tool use sub-command with invalid argument format")
     public void testToolUseSubCommandWithInvalidToolId(String toolId) throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("use", toolId);
-        toolCommand.execute();
+        ToolUseCommand toolUseCommand = new ToolUseCommand(printStream, printStream, false);
+        new CommandLine(toolUseCommand).parseArgs(toolId);
+        toolUseCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replaceAll("\r", ""), getOutput("tool-use-with-invalid-tool-id.txt"));
     }
 
     @Test(description = "Test tool pull sub-command with invalid tool version")
     public void testToolUseSubCommandWithInvalidToolVersion() throws IOException {
-        ToolCommand toolCommand = new ToolCommand(printStream, printStream, false);
-        new CommandLine(toolCommand).parseArgs("use", "tool_id:1.1");
-        toolCommand.execute();
+        ToolUseCommand toolUseCommand = new ToolUseCommand(printStream, printStream, false);
+        new CommandLine(toolUseCommand).parseArgs("tool_id:1.1");
+        toolUseCommand.execute();
         String buildLog = readOutput(true);
         Assert.assertEquals(buildLog.replaceAll("\r", ""), getOutput("tool-use-with-invalid-tool-version.txt"));
     }
