+++ cli/ballerina-cli/src/test/resources/test-resources/command-outputs/unix/tool-pull-with-invalid-tool-id.txt
@@ -1,4 +1,4 @@
 ballerina: invalid tool id.
 
 USAGE:
-    bal tool pull <tool-id>[:<version>]
+    bal tool pull <tool-id>
