+++ cli/ballerina-cli/src/test/resources/test-resources/command-outputs/windows/tool-pull-success.txt
@@ -1,5 +1,2 @@
 WARNING: Specifying a version of the tool is deprecated and may be removed in a future version. Use 'bal tool pull <tool-id>' to pull the latest version of the tool.
-ballerina: invalid tool version. Invalid version: '1.1'. Unexpected character 'EOI(null)' at position '3', expecting '[DOT]'
-
-USAGE:
-    bal tool pull <tool-id>
+tool 'luhee:1.1.0' successfully set as the active version.
