+++ cli/ballerina-cli/src/test/resources/test-resources/command-outputs/windows/tool-pull-with-invalid-tool-version.txt
@@ -1,4 +1,5 @@
+WARNING: Specifying a version of the tool is deprecated and may be removed in a future version. Use 'bal tool pull <tool-id>' to pull the latest version of the tool.
 ballerina: invalid tool version. Invalid version: '1.1'. Unexpected character 'EOI(null)' at position '3', expecting '[DOT]'
 
 USAGE:
-    bal tool pull <tool-id>[:<version>]
+    bal tool pull <tool-id>
