+++ cli/ballerina-cli/src/test/resources/test-resources/command-outputs/windows/tool-pull-with-no-args.txt
@@ -1,4 +1,4 @@
 ballerina: tool id is not provided.
 
 USAGE:
-    bal tool pull <tool-id>[:<version>]
+    bal tool pull <tool-id>
