+++ cli/ballerina-cli/src/test/resources/test-resources/command-outputs/windows/tool-pull-with-too-many-args.txt
@@ -1,4 +1,4 @@
 ballerina: too many arguments.
 
 USAGE:
-    bal tool pull <tool-id>[:<version>]
+    bal tool pull <tool-id>
